requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
scikit-learn
pandas
numpy
langchain
langchain-community
huggingface_hub
apscheduler
python-dotenv
lxml
email-validator
